the storage of the auth token is now handled by a dedicated class
